addressing hasStarted
hasStarted has a very clear (almost didactical) semantic: has
FileWatcher started listening or not? but it is also redundant with the
state of streamRef (it is false if streamRef is nul, true otherwise) so
we end up with tedious "bookkeeping" code.

Simply getting removing hasStarted to replace it with a `streamRef ==
nil` test, would mean loosing the crystal clear / clean semantics.

So.. to solve this and get rid of the ugly bookkeeping, lets rewrite
hasStarted as a computed property over srteamRef - resolves #4